Add list_available_devices for web platform
