+++ .github/workflows/ci.yml
@@ -8,7 +8,8 @@ on:
 
 env:
   python-version: "3.9"
-  node-version: "16"
+  rust-version: 1.66.1
+  node-version: 18.13.0
   poetry-version: 1.2.2
   # TODO: We stick to PostgreSQL 12 for the moment given later versions are
   # much slower (postgresql tests runs in ~9mn on 12 vs ~36mn on 14 !)
