+++ oxidation/client/src/views/TestPage.vue
@@ -53,14 +53,16 @@ import { ref } from 'vue';
 import { libparsec } from '../plugins/libparsec';
 
 const name = ref('Scruffy');
-const message = ref('bonyour!');
 
 async function onSubmit(): Promise<void> {
   console.log('Submitting');
+  await libparsec.testGenDefaultDevices();
+  const devices = await libparsec.listAvailableDevices('PATH/TO/CONFIG/.config/parsec');
+  console.log(devices);
   const key = await libparsec.login('a7d1edf7a8014e76b2d89e63f430d9fc@010cfc8b95ea4436bd594d919ec966fc');
   console.log(key);
-  const rep = await libparsec.loggedCoreGetTestDeviceId(key);
-  console.log(rep);
+  const deviceID = await libparsec.loggedCoreGetTestDeviceId(key);
+  console.log(deviceID);
 }
 </script>
