+++ oxidation/libparsec/Cargo.toml
@@ -17,6 +17,7 @@ libparsec_core = { path = "crates/core", default-features = false }
 libparsec_crypto = { path = "crates/crypto", default-features = false }
 libparsec_client_types = { path = "crates/client_types", default-features = false }
 libparsec_platform_async = { path = "crates/platform_async", default-features = false }
+libparsec_platform_device_loader = { path = "crates/platform_device_loader", default-features = false }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 libparsec_protocol = { path = "crates/protocol", default-features = false }
