+++ oxidation/libparsec/crates/client_types/src/test/test_local_device.rs
@@ -134,10 +134,10 @@ use std::str::FromStr;
     })
 )]
 fn serde_legacy_device_file(#[case] raw: &[u8], #[case] expected: LegacyDeviceFile) {
-    let decoded = LegacyDeviceFile::decode(raw).unwrap();
-    assert_eq!(decoded, expected);
+    let device = LegacyDeviceFile::load(raw).unwrap();
+    assert_eq!(device, expected);
 
     // Roundtrip
-    let roundtrip_raw = decoded.dump();
-    assert_eq!(LegacyDeviceFile::decode(&roundtrip_raw).unwrap(), expected);
+    let roundtrip_raw = device.dump();
+    assert_eq!(LegacyDeviceFile::load(&roundtrip_raw).unwrap(), expected);
 }
