+++ oxidation/libparsec/crates/platform_async/Cargo.toml
@@ -17,7 +17,7 @@ instant = "0.1.12"
 futures-lite = "1.12.0"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen = "0.2.82"
+wasm-bindgen = "0.2.83"
 wasm-bindgen-futures = "0.4.31"
 # Window feature is require for `Timer`.
 web-sys = { version = "0.3.60", features = ["Window"] }
