+++ oxidation/libparsec/crates/platform_device_loader/src/native/mod.rs
@@ -1,0 +1,7 @@
+// Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
+
+mod local_device;
+mod local_device_file;
+
+pub use local_device::*;
+pub use local_device_file::*;
